Add option to create traders on deploy
